added partials for easier styling
all html files are now .ejs files. ejs files are the same as html except you can embed javascript into them. header.ejs and footer.ejs can now be used on any new pages we need to make so that we dont have to rewrite the header and footer for every new page.